requirements.txt
discord_py_interactions==5.12.0
discord_py_slash_command==4.2.1
gspread==6.1.0
